+++ packages/dart_firebase_admin/lib/auth.dart
@@ -1,1 +1,1 @@
-export 'src/auth.dart';
+export 'src/auth.dart' hide UserMetadataToJson;